Fix crash on cleaning empty table

+++ hackernewsbot/dbcleaner.py
@@ -15,12 +15,15 @@ class DatabaseCleaner(object):
         with self._database.cursor() as cursor:
             cursor.execute('SELECT max(ord) from processingStatus;')
             highest_ord, = cursor.fetchone()
-            stale_ord = highest_ord - self._stories_to_keep
-            cursor.execute(('DELETE FROM stories USING processingStatus ' +
-                            'WHERE stories.id = processingStatus.id ' +
-                            'AND processingStatus.ord <= %s;'),
-                           (stale_ord, ))
-            cursor.execute('DELETE FROM processingStatus WHERE ord <= %s;',
-                           (stale_ord, ))
-            logging.debug('deleted stories <= {}'.format(stale_ord))
+            if highest_ord:
+                stale_ord = highest_ord - self._stories_to_keep
+                cursor.execute(('DELETE FROM stories USING processingStatus ' +
+                                'WHERE stories.id = processingStatus.id ' +
+                                'AND processingStatus.ord <= %s;'),
+                               (stale_ord, ))
+                cursor.execute('DELETE FROM processingStatus WHERE ord <= %s;',
+                               (stale_ord, ))
+                logging.debug('deleted stories <= {}'.format(stale_ord))
+            else:
+                logging.debug('no stories to delete')
         self._database.commit()
